+++ src/components/lobby/MatchFoundModal.jsx
@@ -29,7 +29,7 @@ export default function MatchFoundModal({ preMatch }) {
       `mm/match/${preMatch.id}/player-ready/`,
       token
     )
-    if (response && response.errorMsg) {
+    if (response.errorMsg) {
       dispatch(
         addToast({
           content: response.errorMsg,